+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## 0.1.0 (Unreleased)
+## 0.1.0 (August 21, 2017)
 
 - Add brightbox image data source support
 - Update provider to new breakout format
-
+- Make sure terraform refresh removes deleted servers
+- Add import support for servers